+++ src/dkbd/util/tiles.h
@@ -3,9 +3,9 @@
 /**
  * Establece el tile en las coordenadas en PLAN_A
  */
-void tiles_putMapTile(u16 tile, s16 x, s16 y, u8 onTop);
+void tiles_putMapTile(u8 tile, s16 x, s16 y, u8 onTop);
 
 /**
  * Establece un tile en las coordenadas en PLAN_B
  */
-void tiles_putBackgroundTile(u16 tile, s16 x, s16 y, u8 onTop);
+void tiles_putBackgroundTile(u8 tile, s16 x, s16 y, u8 onTop);
